gts7xlwifi: Configure Pogo keyboard as orientation unaware
This way the device is configured with a fixed orientation of 0 (no rotation)
and avoids being mismatched with the applied display rotation of 270.

Fixes dpad keys (no remapping required) and also volume up and down keys,
which were broken when swapped based on screen orientation.

Partial output of `dumpsys input`:

  Device 3: Book Cover Keyboard Slim (EF-DT730)
    EventHub Devices: [ 12 ]
    Generation: 49
    IsExternal: false
    IsWaking: false
    AssociatedDisplayPort: <none>
    AssociatedDisplayUniqueIdByPort: <none>
    AssociatedDisplayUniqueIdByDescriptor: <none>
    HasMic:     false
    Sources: KEYBOARD
    KeyboardType: 2
    ControllerNum: 0
    Keyboard Input Mapper:
      Parameters:
        OrientationAware: false
        HandlesKeyRepeat: false
      Orientation: Rotation0

Change-Id: Ie50994c74391073b8034c2f68cc24513e9c0a08e